# Support TFLite and TFJS conversion for CenterNet multi-class keypoints